GBE-679: More extraction, two clusters of circularity left in remains
