+++ expo/gbe/models/__init__.py
@@ -2,21 +2,13 @@ from user_message import UserMessage
 from available_interest import AvailableInterest
 from remains import (
     Conference,
-    Biddable,
     AudioInfo,
-    LightingInfo,
-    StageInfo,
     TechInfo,
     CueInfo,
-
-    Room,
+    LightingInfo,
+    StageInfo,
     ConferenceDay,
     VolunteerWindow,
-    Event,
-    GenericEvent,
-    AdBid,
-    ArtBid,
-    ClassProposal,
 )
 from show import Show
 from volunteer_interest import VolunteerInterest
@@ -25,7 +17,10 @@ from performer import Performer
 from persona import Persona
 from troupe import Troupe
 from combo import Combo
+from biddable import Biddable
 from act import Act
+from event import Event
+from generic_event import GenericEvent
 from class_model import Class
 from bid_evaluation import BidEvaluation
 from performer_festivals import PerformerFestivals
@@ -34,3 +29,5 @@ from vendor import Vendor
 from costume import Costume
 from conference_volunteer import ConferenceVolunteer
 from profile_preferences import ProfilePreferences
+from class_proposal import ClassProposal
+from room import Room
